requirements.txt

python-telegram-bot==13.15
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
